Changed The csvArrayList class

+++ demo/src/main/java/com/example/MlModel.java
@@ -2,7 +2,7 @@ package com.example;
 import java.util.List;
 
 public interface MlModel {
-    void fit(List<List<Integer>> X_train, List<Integer> y_train);
-    List<Integer> predict(List<List<Integer>> X_test); 
-    public double calculateAccuracy(List<Integer> yTrue, List<Integer> yPred);
+    void fit(List<List<Double>> X_train, List<Double> y_train);
+    List<Double> predict(List<List<Double>> X_test);
+    public double calculateAccuracy(List<Double> yTrue, List<Double> yPred);
 }
